backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
mysql-connector-python==9.5.0
Pillow==12.0.0
gunicorn
whitenoise
python-dotenv